ISSUE-045 - Busy-work, relocating and renaming AngledFlagWorkspace (Workspace, not WorkSpace).
