+++ src/components/Workspaces/CustomWorkSpaces/AngledFlagWorkspace/index.ts
@@ -1,0 +1,1 @@
+export * from "./AngledFlagWorkspace";
